third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/source-frame-count-expected.txt

Tests that scripts panel does not create too many source frames.


Running: testSourceFramesCount
Reloading page...
Page reloaded.
Too many frames opened: 4
Visible view: script5.js
